+++ tests/unit/filters/network.rs
@@ -8,6 +8,8 @@ mod parse_tests {
         hostname: Option<String>,
         opt_domains: Option<Vec<Hash>>,
         opt_not_domains: Option<Vec<Hash>>,
+        opt_to_domains: Option<Vec<Hash>>,
+        opt_not_to_domains: Option<Vec<Hash>>,
         modifier_option: Option<String>,
 
         // filter type
@@ -49,6 +51,8 @@ mod parse_tests {
                 hostname: filter.hostname.as_ref().map(|host| host.to_string()),
                 opt_domains: filter.opt_domains.as_ref().cloned(),
                 opt_not_domains: filter.opt_not_domains.as_ref().cloned(),
+                opt_to_domains: filter.opt_to_domains.as_ref().cloned(),
+                opt_not_to_domains: filter.opt_not_to_domains.as_ref().cloned(),
                 modifier_option: filter
                     .modifier_option
                     .as_ref()
@@ -94,6 +98,8 @@ mod parse_tests {
             hostname: None,
             opt_domains: None,
             opt_not_domains: None,
+            opt_to_domains: None,
+            opt_not_to_domains: None,
             modifier_option: None,
 
             // filter type
@@ -727,6 +733,63 @@ mod parse_tests {
                 Default::default(),
             );
             assert_eq!(filter.err(), Some(NetworkFilterError::NoSupportedDomains));
+        }
+    }
+
+    #[test]
+    fn parses_to() {
+        {
+            let filter =
+                NetworkFilter::parse("*$script,to=bar.com", true, Default::default()).unwrap();
+            assert_eq!(
+                filter.opt_to_domains,
+                Some(vec![utils::fast_hash("bar.com")])
+            );
+            assert_eq!(filter.opt_not_to_domains, None);
+        }
+        {
+            let filter =
+                NetworkFilter::parse("*$script,to=bar.com|baz.com", true, Default::default())
+                    .unwrap();
+            let mut domains = vec![utils::fast_hash("bar.com"), utils::fast_hash("baz.com")];
+            domains.sort_unstable();
+            assert_eq!(filter.opt_to_domains, Some(domains));
+            assert_eq!(filter.opt_not_to_domains, None);
+        }
+        {
+            let filter =
+                NetworkFilter::parse("||it^$3p,to=~example.it", true, Default::default()).unwrap();
+            assert_eq!(filter.opt_to_domains, None);
+            assert_eq!(
+                filter.opt_not_to_domains,
+                Some(vec![utils::fast_hash("example.it")])
+            );
+        }
+        {
+            let filter = NetworkFilter::parse(
+                "*$script,from=beforeitsnews.com,to=gstatic.com",
+                true,
+                Default::default(),
+            )
+            .unwrap();
+            assert_eq!(
+                filter.opt_domains,
+                Some(vec![utils::fast_hash("beforeitsnews.com")])
+            );
+            assert_eq!(
+                filter.opt_to_domains,
+                Some(vec![utils::fast_hash("gstatic.com")])
+            );
+        }
+        {
+            let filter =
+                NetworkFilter::parse(r"*$script,to=/^img[a-z]+\.buzz/", true, Default::default());
+            assert_eq!(filter.err(), Some(NetworkFilterError::NoSupportedDomains));
+        }
+        {
+            let filter = NetworkFilter::parse("||foo.com", true, Default::default()).unwrap();
+            assert_eq!(filter.opt_to_domains, None);
+            assert_eq!(filter.opt_not_to_domains, None);
         }
     }